Assitente_Tecnico_Impressoras/requirements.txt
langchain_ollama
langchain_community
langchain_core
langchain_chroma
unstructured
pdfminer.six
pi_heif
unstructured_inference
pdf2image
pytesseract
unstructured_pytesseract
streamlit
dotenv
langchain_openai